[arcane] Déplace la définition de 'ComponentItemVectorView' et 'ComponentPartItemVectorView' dans 'arcane/core/materials'
